use per-instance temporary directory
The temporary names used as preprocessor output are not unique, so if
another generator instance is running, one can be writing into a file
while the other is reading it. This usually manifests as an error like
`<some_file.idl>:1:0: error: mismatched input '<EOF>' expecting <etc>`.

Fix by randomly generating a unique directory, plus being nice and
cleaning it up if it is unlikely to be useful.

Note that specifying a temporary directory is not actually implemented,
despite a flag to do so being listed in the help.

Signed-off-by: Thomas Watson <twatson52@icloud.com>